fix: move pages to containers and fix imports on next routing
