+++ containers/home.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 import { render, screen } from '@testing-library/react';
-import App from './index';
+import App from '../pages/index';
 
 describe('App', () => {
   it('renders without crashing', () => {
